homework4/my-root-requirements.txt
faiss-cpu==1.10.0
fqdn==1.5.1
google_search_results==2.4.2
isoduration==20.11.0
jupyter==1.1.1
keyring==25.6.0
langchain-community==0.3.20
langchain-openai==0.3.9
libmagic==1.0
numexpr==2.10.2
pipdeptree==2.26.0
ptpython==3.0.29
python-magic-bin==0.4.14
sentence-transformers==3.4.1
streamlit==1.43.2
tinycss2==1.4.0
unstructured==0.17.2
uri-template==1.3.0
webcolors==24.11.1
wikipedia==1.4.0
